separate app from sync
