+++ app/app.go
@@ -3,17 +3,14 @@ package app
 import (
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"path/filepath"
-	"time"
 
 	"github.com/crypto-org-chain/cronos/x/cronos"
 	"github.com/crypto-org-chain/cronos/x/cronos/middleware"
 
 	"github.com/cosmos/cosmos-sdk/client"
-
 	"github.com/cosmos/cosmos-sdk/codec/types"
 	"github.com/cosmos/cosmos-sdk/server"
 	"github.com/gorilla/mux"
@@ -134,7 +131,6 @@ import (
 	cronoskeeper "github.com/crypto-org-chain/cronos/x/cronos/keeper"
 	evmhandlers "github.com/crypto-org-chain/cronos/x/cronos/keeper/evmhandlers"
 	cronostypes "github.com/crypto-org-chain/cronos/x/cronos/types"
-	tmtypes "github.com/tendermint/tendermint/types"
 
 	// unnamed import of statik for swagger UI support
 	_ "github.com/crypto-org-chain/cronos/client/docs/statik"
@@ -311,19 +307,6 @@ type App struct {
 	experimental bool
 }
 
-// Simplify block height for header
-type Header struct {
-	Height int64 `json:"height,omitempty"`
-}
-
-type Block struct {
-	Header Header `json:"header"`
-}
-
-type GetLatestBlockResponse struct {
-	Block *Block `json:"block,omitempty"`
-}
-
 // New returns a reference to an initialized chain.
 // NewSimApp returns a reference to an initialized SimApp.
 func New(
@@ -395,154 +378,14 @@ func New(
 			}
 			versionDB := tmdb.NewStore(plainDB, historyDB, changesetDB)
 			isGrpcOnly := cast.ToBool(appOpts.Get(cronosappclient.FlagIsGrpcOnly))
+			remoteGrpcUrl := cast.ToString(appOpts.Get(cronosappclient.FlagRemoteGrpcUrl))
+			isLocal := cast.ToBool(appOpts.Get(cronosappclient.FlagIsLocal))
+			remoteUrl := cast.ToString(appOpts.Get(cronosappclient.FlagRemoteUrl))
+			remoteWsUrl := cast.ToString(appOpts.Get(cronosappclient.FlagRemoteWsUrl))
+			concurrency := cast.ToInt(appOpts.Get(cronosappclient.FlagConcurrency))
 			if isGrpcOnly {
-				const defaultMaxRetry = 50
-				latestVersion, err := versionDB.GetLatestVersion()
-				fmt.Printf("mm-latestVersion: %+v\n", latestVersion)
-				if err != nil {
-					panic(err)
-				}
-				startBlockNum := latestVersion
-				if startBlockNum < 0 {
-					startBlockNum = 0
-				}
-				nextBlockNum := int(startBlockNum) + 1
-				remoteGrpcUrl := cast.ToString(appOpts.Get(cronosappclient.FlagRemoteGrpcUrl))
-				maxBlockNum := -1
-				for i := 0; i < defaultMaxRetry; i++ {
-					if i > 0 {
-						time.Sleep(time.Second)
-					}
-					resp, err := http.Get(fmt.Sprintf("%s/cosmos/base/tendermint/v1beta1/blocks/latest", remoteGrpcUrl))
-					if err != nil {
-						fmt.Printf("error making http request: %s\n", err)
-						continue
-					}
-
-					bz, err := ioutil.ReadAll(resp.Body)
-					if err != nil {
-						continue
-					}
-					if resp.StatusCode != http.StatusOK {
-						continue
-					}
-					resp.Body.Close()
-					res := new(GetLatestBlockResponse)
-					err = tmjson.Unmarshal(bz, res)
-					if err != nil {
-						fmt.Printf("mm-read-res-err: %+v\n", err)
-						continue
-					}
-					maxBlockNum = int(res.Block.Header.Height)
-					fmt.Printf("mm-maxBlockNum: %d\n", maxBlockNum)
-					break
-				}
-				if maxBlockNum < 0 {
-					panic(fmt.Sprintf("max retries %d reached", defaultMaxRetry))
-				}
-
-				interval := time.Second
-				directory := filepath.Join(rootDir, "data", "file_streamer")
-				// streamer write the file blk by blk with concurrency 1
-				streamer := cronosfile.NewBlockFileWatcher(1, maxBlockNum, func(blockNum int) string {
-					return cronosfile.GetLocalDataFileName(directory, blockNum)
-				}, true)
-				streamer.Start(nextBlockNum, interval)
-				go func() {
-					chData, chErr := streamer.SubscribeData(), streamer.SubscribeError()
-					for {
-						select {
-						case data := <-chData:
-							pairs, err := cronosfile.DecodeData(data.Data)
-							fmt.Printf("mm-pairs: %+v, %+v\n", len(pairs), err)
-							if err == nil {
-								versionDB.PutAtVersion(int64(data.BlockNum), pairs)
-							}
-						case err := <-chErr:
-							// fail read
-							fmt.Println("mm-fail-read-panic")
-							panic(err)
-						}
-					}
-				}()
-
-				isLocal := cast.ToBool(appOpts.Get(cronosappclient.FlagIsLocal))
-				remoteUrl := cast.ToString(appOpts.Get(cronosappclient.FlagRemoteUrl))
-				remoteWsUrl := cast.ToString(appOpts.Get(cronosappclient.FlagRemoteWsUrl))
-				concurrency := cast.ToInt(appOpts.Get(cronosappclient.FlagConcurrency))
-				synchronizer := cronosfile.NewBlockFileWatcher(concurrency, maxBlockNum, func(blockNum int) string {
-					return fmt.Sprintf("%s/%s", remoteUrl, cronosfile.DataFileName(blockNum))
-				}, isLocal)
-				synchronizer.Start(nextBlockNum, interval)
-				go func() {
-					// max retry for temporary io error
-					maxRetry := concurrency * 2
-					retry := 0
-					chData, chErr := synchronizer.SubscribeData(), synchronizer.SubscribeError()
-					for {
-						select {
-						case data := <-chData:
-							file := cronosfile.GetLocalDataFileName(directory, data.BlockNum)
-							fmt.Printf("mm-data.BlockNum: %+v\n", data.BlockNum)
-							if err := os.WriteFile(file, data.Data, 0644); err != nil {
-								fmt.Println("mm-WriteFile-panic")
-								panic(err)
-							}
-							retry = 0
-							fmt.Println("mm-reset-retry")
-							if data.BlockNum > maxBlockNum {
-								streamer.SetMaxBlockNum(data.BlockNum)
-							}
-						case err := <-chErr:
-							retry++
-							fmt.Println("mm-retry", retry)
-							if retry == maxRetry {
-								// data corrupt
-								fmt.Println("mm-data-corrupt-panic")
-								panic(err)
-							}
-						}
-					}
-				}()
-
-				go func() {
-					for i := 0; i < defaultMaxRetry; i++ {
-						if i > 0 {
-							time.Sleep(time.Second)
-						}
-						wsClient := cronosappclient.NewWebsocketClient(remoteWsUrl)
-						chResult, err := wsClient.Subscribe()
-						if err != nil {
-							fmt.Printf("mm-subscribed[%+v]: %+v\n", i, err)
-							continue
-						}
-						fmt.Println("subscribing")
-						err = wsClient.Send("subscribe", []string{
-							"tm.event='NewBlockHeader'",
-						})
-						if err != nil {
-							fmt.Printf("mm-subscribed: %+v\n", err)
-							continue
-						}
-						i = 0
-						fmt.Println("subscribed ws")
-						for res := range chResult {
-							if res == nil || res.Data == nil {
-								continue
-							}
-							data, ok := res.Data.(tmtypes.EventDataNewBlockHeader)
-							if !ok {
-								continue
-							}
-							blockNum := int(data.Header.Height)
-							fmt.Printf("mm-set-max-blk: %+v\n", blockNum)
-							synchronizer.SetMaxBlockNum(blockNum)
-						}
-					}
-					panic(fmt.Sprintf("max retries %d reached", defaultMaxRetry))
-				}()
+				cronosfile.Sync(versionDB, remoteGrpcUrl, remoteUrl, remoteWsUrl, rootDir, isLocal, concurrency)
 			}
-
 			// default to exposing all
 			exposeStoreKeys := make([]storetypes.StoreKey, 0, len(keys))
 			for _, storeKey := range keys {
